+++ codex-rs/tui/src/exec_cell/mod.rs
@@ -1,0 +1,12 @@
+mod model;
+mod render;
+
+pub(crate) use model::CommandOutput;
+#[cfg(test)]
+pub(crate) use model::ExecCall;
+pub(crate) use model::ExecCell;
+pub(crate) use render::OutputLinesParams;
+pub(crate) use render::TOOL_CALL_MAX_LINES;
+pub(crate) use render::new_active_exec_command;
+pub(crate) use render::output_lines;
+pub(crate) use render::spinner;
